Add private queue to synchronize access to stub in URLProtocol stub
