+++ EssentialDevelopper/EssentialDeveloperFrameworkTests/Feed API/URLSessionHTTPClientTests.swift
@@ -158,68 +158,3 @@ private extension URLSessionHTTPClientTests {
         return NSError(domain: "TestHTTPClientError", code: 1, userInfo: nil)
     }
 }
-// MARK: -
-private extension URLSessionHTTPClientTests {
-    class URLProtocolStub: URLProtocol {
-        var requestedURL: URL?
-        
-        public static var stub: Stub?
-        
-        private typealias ObserverRequestBlock = (URLRequest) -> Void
-        private static var observeRequestBlock: ObserverRequestBlock?
-        
-        struct Stub {
-            var data: Data?
-            var response: URLResponse?
-            var error: Error?
-        }
-        
-        public static func stub(data: Data?, response: URLResponse?, error: Error?) {
-            stub = Stub(data: data, response: response, error: error)
-        }
-        
-        static func startInterceptingRequests() {
-            URLProtocol.registerClass(URLProtocolStub.self)
-        }
-        
-        static func stopInterceptingRequests() {
-            URLProtocol.unregisterClass(URLProtocolStub.self)
-            stub = nil
-            observeRequestBlock = nil
-        }
-        
-        override class func canInit(with request: URLRequest) -> Bool {
-            return true
-        }
-        
-        override class func canonicalRequest(for request: URLRequest) -> URLRequest {
-            return request
-        }
-        
-        override func startLoading() {
-            if let requestObserver = Self.observeRequestBlock {
-                client?.urlProtocolDidFinishLoading(self)
-                return requestObserver(self.request)
-            }
-            if let data = URLProtocolStub.stub?.data {
-                client?.urlProtocol(self, didLoad: data)
-            }
-            
-            if let response = URLProtocolStub.stub?.response {
-                client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
-            }
-            
-            if let error = URLProtocolStub.stub?.error {
-                client?.urlProtocol(self, didFailWithError: error)
-            }
-            
-            client?.urlProtocolDidFinishLoading(self)
-        }
-        
-        override func stopLoading() {}
-        
-        static func observeRequest(completion: @escaping (URLRequest) -> ()) {
-            URLProtocolStub.observeRequestBlock = completion
-        }
-    }
-}
